refactor: remove `IRemoteEnvironment` interface which is useless
