+++ antareslauncher/use_cases/check_remote_queue/check_queue_controller.py
@@ -1,9 +1,7 @@
 from dataclasses import dataclass
 
 from antareslauncher.data_repo.idata_repo import IDataRepo
-from antareslauncher.use_cases.check_remote_queue.slurm_queue_show import (
-    SlurmQueueShow,
-)
+from antareslauncher.use_cases.check_remote_queue.slurm_queue_show import SlurmQueueShow
 from antareslauncher.use_cases.retrieve.state_updater import StateUpdater
