+++ antareslauncher/use_cases/kill_job/job_kill_controller.py
@@ -2,14 +2,14 @@ from dataclasses import dataclass
 
 from antareslauncher.data_repo.idata_repo import IDataRepo
 from antareslauncher.display.idisplay import IDisplay
-from antareslauncher.remote_environnement.iremote_environment import (
-    IRemoteEnvironment,
+from antareslauncher.remote_environnement.remote_environment_with_slurm import (
+    RemoteEnvironmentWithSlurm,
 )
 
 
 @dataclass
 class JobKillController:
-    env: IRemoteEnvironment
+    env: RemoteEnvironmentWithSlurm
     display: IDisplay
     repo: IDataRepo
