+++ antareslauncher/use_cases/retrieve/retrieve_controller.py
@@ -2,16 +2,12 @@ from antareslauncher.data_repo.data_reporter import DataReporter
 from antareslauncher.data_repo.idata_repo import IDataRepo
 from antareslauncher.display.idisplay import IDisplay
 from antareslauncher.file_manager.file_manager import FileManager
-from antareslauncher.remote_environnement import iremote_environment
-from antareslauncher.use_cases.retrieve.clean_remote_server import (
-    RemoteServerCleaner,
+from antareslauncher.remote_environnement.remote_environment_with_slurm import (
+    RemoteEnvironmentWithSlurm,
 )
-from antareslauncher.use_cases.retrieve.download_final_zip import (
-    FinalZipDownloader,
-)
-from antareslauncher.use_cases.retrieve.final_zip_extractor import (
-    FinalZipExtractor,
-)
+from antareslauncher.use_cases.retrieve.clean_remote_server import RemoteServerCleaner
+from antareslauncher.use_cases.retrieve.download_final_zip import FinalZipDownloader
+from antareslauncher.use_cases.retrieve.final_zip_extractor import FinalZipExtractor
 from antareslauncher.use_cases.retrieve.log_downloader import LogDownloader
 from antareslauncher.use_cases.retrieve.state_updater import StateUpdater
 from antareslauncher.use_cases.retrieve.study_retriever import StudyRetriever
@@ -21,7 +17,7 @@ class RetrieveController:
     def __init__(
         self,
         repo: IDataRepo,
-        env: iremote_environment.IRemoteEnvironment,
+        env: RemoteEnvironmentWithSlurm,
         file_manager: FileManager,
         display: IDisplay,
         state_updater: StateUpdater,
