+++ antareslauncher/use_cases/retrieve/state_updater.py
@@ -2,17 +2,18 @@ from pathlib import Path
 from typing import List
 
 from antareslauncher.display.idisplay import IDisplay
-from antareslauncher.remote_environnement import iremote_environment
+from antareslauncher.remote_environnement.remote_environment_with_slurm import (
+    RemoteEnvironmentWithSlurm,
+)
 from antareslauncher.study_dto import StudyDTO
 
 
 class StateUpdater:
     def __init__(
         self,
-        env: iremote_environment.IRemoteEnvironment,
+        env: RemoteEnvironmentWithSlurm,
         display: IDisplay,
     ):
-
         self._env = env
         self._display = display
         self._current_study: StudyDTO = None
