+++ tests/unit/launcher/test_submitter.py
@@ -5,8 +5,8 @@ import pytest
 
 import antareslauncher.use_cases
 from antareslauncher.display.idisplay import IDisplay
-from antareslauncher.remote_environnement.iremote_environment import (
-    IRemoteEnvironment,
+from antareslauncher.remote_environnement.remote_environment_with_slurm import (
+    RemoteEnvironmentWithSlurm,
 )
 from antareslauncher.study_dto import StudyDTO
 from antareslauncher.use_cases.launch.study_submitter import StudySubmitter
@@ -14,7 +14,7 @@ from antareslauncher.use_cases.launch.study_submitter import StudySubmitter
 
 class TestStudySubmitter:
     def setup_method(self):
-        self.remote_env = mock.Mock(spec_set=IRemoteEnvironment)
+        self.remote_env = mock.Mock(spec_set=RemoteEnvironmentWithSlurm)
         self.display_mock = mock.Mock(spec_set=IDisplay)
         self.study_submitter = StudySubmitter(self.remote_env, self.display_mock)
