+++ tests/unit/retriever/test_retrieve_controller.py
@@ -5,18 +5,19 @@ from unittest.mock import call
 import pytest
 
 import antareslauncher
+import antareslauncher.remote_environnement.remote_environment_with_slurm
 from antareslauncher.display.idisplay import IDisplay
-from antareslauncher.remote_environnement import iremote_environment
-from antareslauncher.study_dto import StudyDTO
-from antareslauncher.use_cases.retrieve.retrieve_controller import (
-    RetrieveController,
+from antareslauncher.remote_environnement.remote_environment_with_slurm import (
+    RemoteEnvironmentWithSlurm,
 )
+from antareslauncher.study_dto import StudyDTO
+from antareslauncher.use_cases.retrieve.retrieve_controller import RetrieveController
 from antareslauncher.use_cases.retrieve.state_updater import StateUpdater
 
 
 class TestRetrieveController:
     def setup_method(self):
-        self.remote_env_mock = mock.Mock(spec=iremote_environment.IRemoteEnvironment)
+        self.remote_env_mock = mock.Mock(spec=RemoteEnvironmentWithSlurm)
         self.file_manager = mock.Mock()
         self.data_repo = mock.Mock()
         self.display = mock.Mock()
