+++ tests/unit/retriever/test_server_cleaner.py
@@ -4,8 +4,11 @@ from unittest import mock
 
 import pytest
 
+import antareslauncher.remote_environnement.remote_environment_with_slurm
 from antareslauncher.display.idisplay import IDisplay
-from antareslauncher.remote_environnement import iremote_environment
+from antareslauncher.remote_environnement.remote_environment_with_slurm import (
+    RemoteEnvironmentWithSlurm,
+)
 from antareslauncher.study_dto import StudyDTO
 from antareslauncher.use_cases.retrieve.clean_remote_server import (
     RemoteServerCleaner,
@@ -15,7 +18,7 @@ from antareslauncher.use_cases.retrieve.clean_remote_server import (
 
 class TestServerCleaner:
     def setup_method(self):
-        self.remote_env_mock = mock.Mock(spec=iremote_environment.IRemoteEnvironment)
+        self.remote_env_mock = mock.Mock(spec=RemoteEnvironmentWithSlurm)
         self.display_mock = mock.Mock(spec_set=IDisplay)
         self.remote_server_cleaner = RemoteServerCleaner(
             self.remote_env_mock, self.display_mock
