+++ tests/unit/retriever/test_study_retriever.py
@@ -8,19 +8,13 @@ from antareslauncher.data_repo.data_reporter import DataReporter
 from antareslauncher.data_repo.idata_repo import IDataRepo
 from antareslauncher.display.idisplay import IDisplay
 from antareslauncher.file_manager.file_manager import FileManager
-from antareslauncher.remote_environnement.iremote_environment import (
-    IRemoteEnvironment,
+from antareslauncher.remote_environnement.remote_environment_with_slurm import (
+    RemoteEnvironmentWithSlurm,
 )
 from antareslauncher.study_dto import StudyDTO
-from antareslauncher.use_cases.retrieve.clean_remote_server import (
-    RemoteServerCleaner,
-)
-from antareslauncher.use_cases.retrieve.download_final_zip import (
-    FinalZipDownloader,
-)
-from antareslauncher.use_cases.retrieve.final_zip_extractor import (
-    FinalZipExtractor,
-)
+from antareslauncher.use_cases.retrieve.clean_remote_server import RemoteServerCleaner
+from antareslauncher.use_cases.retrieve.download_final_zip import FinalZipDownloader
+from antareslauncher.use_cases.retrieve.final_zip_extractor import FinalZipExtractor
 from antareslauncher.use_cases.retrieve.log_downloader import LogDownloader
 from antareslauncher.use_cases.retrieve.state_updater import StateUpdater
 from antareslauncher.use_cases.retrieve.study_retriever import StudyRetriever
@@ -28,7 +22,7 @@ from antareslauncher.use_cases.retrieve.study_retriever import StudyRetriever
 
 class TestStudyRetriever:
     def setup_method(self):
-        env = mock.Mock(spec_set=IRemoteEnvironment)
+        env = mock.Mock(spec_set=RemoteEnvironmentWithSlurm)
         display = mock.Mock(spec_set=IDisplay)
         file_manager = mock.Mock(spec_set=FileManager)
         repo = mock.Mock(spec_set=IDataRepo)
